backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-telegram-bot==20.7
sqlalchemy==2.0.25
alembic==1.13.1

# Database
psycopg2-binary==2.9.9

# PDF Parsing
pdfplumber==0.10.3
tabula-py==2.9.0
pandas==2.1.4

# ML
scikit-learn==1.4.0
numpy==1.26.3
joblib==1.3.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
PyYAML==6.0.1
google-api-python-client==2.169.0
google-auth==2.40.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
